ai_service_chat/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.40
httpx>=0.26.0
pydantic>=2.0.0
